toMatchSnapshot tests
