+++ e2e/__tests__/toMatchSnapshotDeep.test.ts
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import * as path from 'path';
+import {cleanup, makeTemplate, writeFiles} from '../Utils';
+import runJest from '../runJest';
+
+const DIR = path.resolve(__dirname, '../to-match-snapshot');
+const TESTS_DIR = path.resolve(DIR, '__tests__');
+
+beforeEach(() => cleanup(TESTS_DIR));
+afterAll(() => cleanup(TESTS_DIR));
+
+test('handles property matchers with deep properties', () => {
+  const filename = 'handle-property-matchers-with-name.test.js';
+  const template =
+    makeTemplate(`test('handles property matchers with deep properties', () => {
+      expect({ user: { createdAt: $1, name: $2 }}).toMatchSnapshot({ user: { createdAt: expect.any(Date), name: $2 }});
+    });
+    `);
+
+  {
+    writeFiles(TESTS_DIR, {[filename]: template(['new Date()', '"Jest"'])});
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('1 snapshot written from 1 test suite.');
+    expect(exitCode).toBe(0);
+  }
+
+  {
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('Snapshots:   1 passed, 1 total');
+    expect(exitCode).toBe(0);
+  }
+
+  {
+    writeFiles(TESTS_DIR, {[filename]: template(['"string"', '"Jest"'])});
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch(
+      'Snapshot name: `handles property matchers with deep properties 1`',
+    );
+    expect(stderr).toMatch('Expected properties');
+    expect(stderr).toMatch('Snapshots:   1 failed, 1 total');
+    expect(exitCode).toBe(1);
+  }
+
+  {
+    writeFiles(TESTS_DIR, {[filename]: template(['new Date()', '"CHANGED"'])});
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch(
+      'Snapshot name: `handles property matchers with deep properties 1`',
+    );
+    expect(stderr).toMatch('Snapshots:   1 failed, 1 total');
+    expect(exitCode).toBe(1);
+  }
+});
