+++ e2e/__tests__/toMatchSnapshotError.test.ts
@@ -1,0 +1,49 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import * as path from 'path';
+import {cleanup, makeTemplate, writeFiles} from '../Utils';
+import runJest from '../runJest';
+
+const DIR = path.resolve(__dirname, '../to-match-snapshot');
+const TESTS_DIR = path.resolve(DIR, '__tests__');
+
+beforeEach(() => cleanup(TESTS_DIR));
+afterAll(() => cleanup(TESTS_DIR));
+
+
+test('error thrown before snapshot', () => {
+  const filename = 'error-thrown-before-snapshot.test.js';
+  const template = makeTemplate(`test('snapshots', () => {
+      expect($1).toBeTruthy();
+      expect($2).toMatchSnapshot();
+    });`);
+
+  {
+    writeFiles(TESTS_DIR, {
+      [filename]: template(['true', '{a: "original"}']),
+    });
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('1 snapshot written from 1 test suite.');
+    expect(exitCode).toBe(0);
+  }
+
+  {
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('Snapshots:   1 passed, 1 total');
+    expect(exitCode).toBe(0);
+  }
+
+  {
+    writeFiles(TESTS_DIR, {
+      [filename]: template(['false', '{a: "original"}']),
+    });
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).not.toMatch('1 obsolete snapshot found');
+    expect(exitCode).toBe(1);
+  }
+});
